third_party/WebKit/LayoutTests/editing/execCommand/insertHTML-aborted.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<iframe src="javascript:onunload = function() {
    if (!top.container) {
        top.container = document.createElement('div');
        top.container.appendChild(frameElement.parentNode);
    }
}">
</iframe>
<script>
window.onload = function() {
    document.designMode = 'on';
    test(function() {
        document.execCommand('selectall');
        assert_false(document.execCommand('InsertHTML', false, '<p>para</p>'));
    }, 'InsertHTML command should return false if it failed.');
};
</script>
